Move `chia/types/spend_bundle.py` to wallet.
